In progress merging FSC code with newest main branch
