+++ features/__init__.py
@@ -1,7 +1,7 @@
 # developer: taoshi-mbrown
 # Copyright © 2024 Taoshi, LLC
 from .feature_id import FeatureID
-from .feature_source import FeatureCompaction, FeatureSource
+from .feature_source import FeatureCompaction, FeatureSource, get_feature_ids
 from .feature_scaler import FeatureScaler, IndividualScaler, GroupScaler
 from .feature_aggregator import FeatureAggregator, IndividualAggregator, GroupAggregator
 from .feature_collector import FeatureCollector
