+++ meta/meta.json
@@ -1,4 +1,4 @@
 {
-  "subnet_version": "4.4.2",
-  "mining_model_version": "4.0.0"
+  "subnet_version": "5.2.0",
+  "mining_model_version": "5.0.0"
 }
