+++ miner_config.py
@@ -4,5 +4,6 @@
 
 class MinerConfig:
     MIN_VALI_SIZE = 0
-    ACCEPTABLE_INTERVALS = [0, 30]
+    ACCEPTABLE_INTERVALS_PREDICTIONS = [1, 2, 31, 32]
+    ACCEPTABLE_INTERVALS_HASH = [0, 30]
     STD_LOOKBACK = 3
