+++ mining_objects/__init__.py
@@ -1,0 +1,1 @@
+from base_mining_model import BaseMiningModel
