+++ neurons/miner.py
@@ -1,40 +1,48 @@
 # The MIT License (MIT)
 # Copyright © 2023 Yuma Rao
-# developer: Taoshidev
-# Copyright © 2023 Taoshi Inc
-
-
-# Step 1: Import necessary libraries and modules
-import datetime
+# developer: taoshi-mbrown
+# Copyright © 2024 Taoshi, LLC
+import argparse
+import bittensor as bt
+import copy
+from features import FeatureCollector, FeatureSource
+from hashing_utils import HashingUtils
+from miner_config import MinerConfig
+from mining_objects import BaseMiningModel
+from mining_objects.streams.btcusd_5m import (
+    INTERVAL_MS,
+    model_feature_ids,
+    model_feature_scaler,
+    model_feature_sources,
+    prediction_feature_ids,
+    PREDICTION_COUNT,
+    PREDICTION_LENGTH,
+    SAMPLE_COUNT,
+)
+import numpy as np
 import os
 import sys
+import template
 import threading
 import time
-from typing import Tuple
-
-import numpy as np
-
-import template
-import argparse
+from time_util import datetime
 import traceback
-import bittensor as bt
-
-from data_generator.data_generator_handler import DataGeneratorHandler
-from miner_config import MinerConfig
-from mining_objects.base_mining_model import BaseMiningModel
-from mining_objects.mining_utils import MiningUtils
-from time_util.time_util import TimeUtil
 from vali_config import ValiConfig
 from vali_objects.dataclasses.stream_prediction import StreamPrediction
 from vali_objects.request_templates import RequestTemplates
-from vali_objects.utils.vali_utils import ValiUtils
+
+FEATURE_COLLECTOR_TIMEOUT = 10.0
 
 base_mining_model = None
 base_model_id = None
 
+# Cached miner predictions
+miner_preds = {}
+sent_preds = {}
+
 
 def get_config():
-    # Step 2: Set up the configuration parser
+    # Set up the configuration parser
     # This function initializes the necessary command-line arguments.
     # Using command-line arguments allows users to customize various miner settings.
     parser = argparse.ArgumentParser()
@@ -44,7 +52,12 @@ def get_config():
     )
     # Adds override arguments for network and netuid.
     parser.add_argument("--netuid", type=int, default=1, help="The chain subnet uid.")
-    parser.add_argument("--base_model", type=str, default="model_v4_1", help="Choose the base model you want to run (if youre not using a custom one).")
+    parser.add_argument(
+        "--base_model",
+        type=str,
+        default="model_v5_1",
+        help="Choose the base model you want to run (if youre not using a custom one).",
+    )
     # Adds subtensor specific arguments i.e. --subtensor.chain_endpoint ... --subtensor.network ...
     bt.subtensor.add_args(parser)
     # Adds logging specific arguments i.e. --logging.debug ..., --logging.trace .. or --logging.logging_dir ...
@@ -57,7 +70,7 @@ def get_config():
     # To print help message, run python3 template/miner.py --help
     config = bt.config(parser)
 
-    # Step 3: Set up logging directory
+    # Set up logging directory
     # Logging captures events for diagnosis or understanding miner's behavior.
     config.full_path = os.path.expanduser(
         "{}/{}/{}/netuid{}/{}".format(
@@ -74,42 +87,61 @@ def get_config():
     return config
 
 
-def update_predictions(stream_predictions, data_generator_handler, model_chosen):
+def update_predictions(
+    stream_predictions: list[StreamPrediction],
+    feature_source: FeatureSource,
+    model: BaseMiningModel,
+):
     while True:
-        current_time = datetime.datetime.now()
+        current_time = datetime.now()
         if current_time.second < 15:
             bt.logging.debug(f"running update of predictions [{current_time}]")
-            # for each template
+
             for stream_prediction in stream_predictions:
                 if stream_prediction.stream_type not in miner_preds:
                     bt.logging.info(
-                        f"stream type doesn't exist, setting to an empty list for [{stream_prediction.stream_type}]")
+                        f"stream type doesn't exist, setting to an empty list for [{stream_prediction.stream_type}]"
+                    )
                     miner_preds[stream_prediction.stream_type] = []
                 bt.logging.debug(
-                    f"current predicted closes in memory [{miner_preds[stream_prediction.stream_type]}]")
-                bt.logging.info(f"setting predictions for [{stream_prediction.stream_type}]")
-                # get lookback data
-                ts_ranges = TimeUtil.convert_range_timestamps_to_millis(
-                    TimeUtil.generate_range_timestamps(
-                        TimeUtil.generate_start_timestamp(MinerConfig.STD_LOOKBACK), MinerConfig.STD_LOOKBACK))
-                ds = ValiUtils.get_standardized_ds()
-                for ts_range in ts_ranges:
-                    data_generator_handler.data_generator_handler(stream_prediction.topic_id,
-                                                                  0,
-                                                                  stream_prediction.additional_details,
-                                                                  ds,
-                                                                  ts_range)
-                np_ds = np.array(ds)
+                    f"current predicted closes in memory [{miner_preds[stream_prediction.stream_type]}]"
+                )
+                bt.logging.info(
+                    f"setting predictions for [{stream_prediction.stream_type}]"
+                )
 
-                # generate stream predictions
-                predicted_closes = MiningUtils.open_model_prediction_generation(np_ds, model_chosen,
-                                                                                stream_prediction.prediction_size)
-                bt.logging.debug(f"predicted closes [{predicted_closes}]")
+                start_time_ms = current_time.timestamp_ms() - (
+                    SAMPLE_COUNT * INTERVAL_MS
+                )
+
+                feature_samples = feature_source.get_feature_samples(
+                    start_time_ms, INTERVAL_MS, SAMPLE_COUNT
+                )
+
+                # FeatureScaler is not thread safe, so make a local copy
+                feature_scaler = copy.deepcopy(model_feature_scaler)
+                feature_scaler.scale_feature_samples(feature_samples)
+
+                model_input = feature_source.feature_samples_to_array(feature_samples)
+                predictions = model.predict(model_input)
+
+                prediction_samples = feature_source.array_to_feature_samples(
+                    predictions, prediction_feature_ids
+                )
+                feature_scaler.unscale_feature_samples(prediction_samples)
+
+                prediction_array = feature_source.feature_samples_to_array(
+                    prediction_samples, prediction_feature_ids
+                )
+
+                bt.logging.debug(f"predicted closes [{prediction_array}]")
 
                 # set preds in memory
-                miner_preds[stream_prediction.stream_type] = predicted_closes
-                bt.logging.info(f"done setting predictions for [{stream_prediction.stream_type}] in memory "
-                                f"with length [{len(predicted_closes)}]")
+                miner_preds[stream_prediction.stream_type] = prediction_array
+                bt.logging.info(
+                    f"done setting predictions for [{stream_prediction.stream_type}] in memory "
+                    f"with length [{len(prediction_array)}]"
+                )
             time.sleep(15)
 
 
@@ -117,7 +149,7 @@ def get_model_dir(model):
     return ValiConfig.BASE_DIR + model
 
 
-def is_invalid_validator(metagraph, hotkey):
+def is_invalid_validator(metagraph, hotkey, acceptable_intervals):
     """
     - step 1: check to see if the vali is in the metagraph
     - step 2: check to see if the vali has min threshold
@@ -126,10 +158,10 @@ def is_invalid_validator(metagraph, hotkey):
 
     # step 1 - check to see if the vali is in the metagraph
     if hotkey not in metagraph.hotkeys:
-        bt.logging.trace(f'Hotkey does not exist in metagraph [{hotkey}]')
+        bt.logging.trace(f"Hotkey does not exist in metagraph [{hotkey}]")
         return True
 
-    bt.logging.info(f'Valid hotkey [{hotkey}]')
+    bt.logging.info(f"Valid hotkey [{hotkey}]")
 
     # step 2: check to see if the vali has min threshold
     uid = None
@@ -142,16 +174,18 @@ def is_invalid_validator(metagraph, hotkey):
     bt.logging.debug(f"stake of [{hotkey}]: [{stake}]")
 
     if stake < MinerConfig.MIN_VALI_SIZE:
-        bt.logging.info(f"Denied due to low stake. Min threshold [{MinerConfig.MIN_VALI_SIZE}]")
+        bt.logging.info(
+            f"Denied due to low stake. Min threshold [{MinerConfig.MIN_VALI_SIZE}]"
+        )
         return True
 
-    # ADDING WITH V5.0.0
+    # ADDING WITH V5.1.0
     # step 3: ensure the request is in the required time window
     # current_time = datetime.datetime.now()
     #
-    # bt.logging.debug(f"Acceptable intervals for requests [{MinerConfig.ACCEPTABLE_INTERVALS}]")
+    # bt.logging.debug(f"Acceptable intervals for requests [{acceptable_intervals}]")
     #
-    # if current_time.minute not in MinerConfig.ACCEPTABLE_INTERVALS:
+    # if current_time.minute not in acceptable_intervals:
     #     bt.logging.info(f"Denied due to incorrect interval [{current_time.minute}m]")
     #     return True
 
@@ -159,49 +193,12 @@ def is_invalid_validator(metagraph, hotkey):
 
 
 # Main takes the config and starts the miner.
-def main( config ):
+def main(config):
     base_mining_models = {
-        "model_v4_1": {
-            "model_dir": get_model_dir("/mining_models/model_v4_1.h5"),
-            "window_size": 100,
-            "id": "model2308",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
-        },
-        "model_v4_2": {
-            "model_dir": get_model_dir("/mining_models/model_v4_2.h5"),
-            "window_size": 500,
-            "id": "model3005",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
-        },
-        "model_v4_3": {
-            "model_dir": get_model_dir("/mining_models/model_v4_3.h5"),
-            "window_size": 100,
-            "id": "model3103",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
-        },
-        "model_v4_4": {
-            "model_dir": get_model_dir("/mining_models/model_v4_4.h5"),
-            "window_size": 100,
-            "id": "model3104",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
-        },
-        "model_v4_5": {
-            "model_dir": get_model_dir("/mining_models/model_v4_5.h5"),
-            "window_size": 100,
-            "id": "model3105",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
-        },
-        "model_v4_6": {
-            "model_dir": get_model_dir("/mining_models/model_v4_6.h5"),
-            "window_size": 100,
-            "id": "model3106",
-            "features": BaseMiningModel.base_model_dataset,
-            "rows": 601
+        "model_v5_1": {
+            "creation_id": "model5001",
+            "filename": "/mining_models/model_v5_1.h5",
+            "sample_count": 100,
         },
     }
 
@@ -214,9 +211,7 @@ def main( config ):
     global base_mining_model
     global base_model_id
     global miner_preds
-
-    # where we'll store miner predictions in memory and reference
-    miner_preds = {}
+    global sent_preds
 
     base_mining_model = None
     base_model_id = config.base_model
@@ -226,11 +221,18 @@ def main( config ):
         bt.logging.debug(f"using an existing base model [{config.base_model}]")
 
         model_chosen = base_mining_models[base_model_id]
+        model_filename = get_model_dir(model_chosen["filename"])
 
-        base_mining_model = BaseMiningModel(4) \
-            .set_window_size(model_chosen["window_size"]) \
-            .set_model_dir(model_chosen["model_dir"]) \
-            .load_model()
+        base_mining_model = BaseMiningModel(
+            filename=model_filename,
+            mode="r",
+            feature_count=len(model_feature_ids),
+            sample_count=model_chosen["sample_count"],
+            prediction_feature_count=len(prediction_feature_ids),
+            prediction_count=PREDICTION_COUNT,
+            prediction_length=PREDICTION_LENGTH,
+        )
+
     else:
         bt.logging.debug("base model not chosen.")
 
@@ -263,7 +265,7 @@ def main( config ):
     my_subnet_uid = metagraph.hotkeys.index(wallet.hotkey.ss58_address)
     bt.logging.info(f"Running miner on uid: {my_subnet_uid}")
 
-    # def tf_blacklist_fn(synapse: template.protocol.TrainingForward) -> Tuple[bool, str]:
+    # def tf_blacklist_fn(synapse: template.protocol.TrainingForward) -> tuple[bool, str]:
     #     # standardizing not accepting tf and tb for now
     #     return False, synapse.dendrite.hotkey
     #
@@ -280,7 +282,7 @@ def main( config ):
     #     bt.logging.debug(f'sending tf with length {len(predictions)}')
     #     return synapse
     #
-    # def tb_blacklist_fn( synapse: template.protocol.TrainingBackward ) -> Tuple[bool, str]:
+    # def tb_blacklist_fn( synapse: template.protocol.TrainingBackward ) -> tuple[bool, str]:
     #     # standardizing not accepting tf and tb for now
     #     return False, synapse.dendrite.hotkey
     #
@@ -295,34 +297,93 @@ def main( config ):
     #     synapse.received = True
     #     return synapse
 
-    def lf_blacklist_fn(synapse: template.protocol.LiveForward) -> Tuple[bool, str]:
-        _is_invalid_validator = is_invalid_validator(metagraph, synapse.dendrite.hotkey)
+    def lf_hash_blacklist_fn(
+        synapse: template.protocol.LiveForwardHash,
+    ) -> tuple[bool, str]:
+        _is_invalid_validator = is_invalid_validator(
+            metagraph, synapse.dendrite.hotkey, MinerConfig.ACCEPTABLE_INTERVALS_HASH
+        )
         return _is_invalid_validator, synapse.dendrite.hotkey
 
-    def lf_priority_fn(synapse: template.protocol.LiveForward) -> float:
-        caller_uid = metagraph.hotkeys.index( synapse.dendrite.hotkey )
-        priority = float( metagraph.S[ caller_uid ] )
-        bt.logging.trace(f'Prioritizing {synapse.dendrite.hotkey} with value: ', priority)
+    def lf_hash_priority_fn(synapse: template.protocol.LiveForwardHash) -> float:
+        caller_uid = metagraph.hotkeys.index(synapse.dendrite.hotkey)
+        priority = float(metagraph.S[caller_uid])
+        bt.logging.trace(
+            f"Prioritizing {synapse.dendrite.hotkey} with value: ", priority
+        )
         return priority
 
-    def live_f(synapse: template.protocol.LiveForward) -> template.protocol.LiveForward:
-
-        bt.logging.debug(f"received lf request on stream type [{synapse.stream_id}]")
-
+    def live_hash_f(
+        synapse: template.protocol.LiveForwardHash,
+    ) -> template.protocol.LiveForwardHash:
+        bt.logging.debug(
+            f"received lf hash request on stream type [{synapse.stream_id}] "
+            f"by vali [{synapse.dendrite.hotkey}]"
+        )
         # Convert the string back to a list using literal_eval
         try:
             if synapse.stream_id in miner_preds:
+                if synapse.stream_id not in sent_preds:
+                    sent_preds[synapse.stream_id] = {}
                 stream_preds = miner_preds[synapse.stream_id]
-                synapse.predictions = bt.tensor(np.array(stream_preds))
-                bt.logging.debug(f"sending lf [{stream_preds}]")
-                bt.logging.debug(f'sending lf with length [{len(stream_preds)}]')
+                sent_preds[synapse.stream_id][synapse.dendrite.hotkey] = stream_preds
+                hashed_preds = HashingUtils.hash_predictions(
+                    wallet.hotkey.ss58_address, str(stream_preds)
+                )
+
+                synapse.hashed_predictions = hashed_preds
+                bt.logging.debug(f"sending hash lf [{stream_preds}]")
+                bt.logging.debug(f"sending hash lf with length [{len(stream_preds)}]")
                 return synapse
             else:
                 bt.logging.error(f"miner preds not stored properly in memory")
         except Exception as e:
             bt.logging.error(f"error returning synapse to vali: {e}")
 
-    # def lb_blacklist_fn(synapse: template.protocol.LiveBackward) -> Tuple[bool, str]:
+    def lf_blacklist_fn(synapse: template.protocol.LiveForward) -> tuple[bool, str]:
+        _is_invalid_validator = is_invalid_validator(
+            metagraph,
+            synapse.dendrite.hotkey,
+            MinerConfig.ACCEPTABLE_INTERVALS_PREDICTIONS,
+        )
+        return _is_invalid_validator, synapse.dendrite.hotkey
+
+    def lf_priority_fn(synapse: template.protocol.LiveForward) -> float:
+        caller_uid = metagraph.hotkeys.index(synapse.dendrite.hotkey)
+        priority = float(metagraph.S[caller_uid])
+        bt.logging.trace(
+            f"Prioritizing {synapse.dendrite.hotkey} with value: ", priority
+        )
+        return priority
+
+    def live_f(synapse: template.protocol.LiveForward) -> template.protocol.LiveForward:
+        bt.logging.debug(
+            f"received lf request on stream type [{synapse.stream_id}] "
+            f"by vali [{synapse.dendrite.hotkey}]"
+        )
+
+        # need to determine if the validator has correctly updated to the latest logic. This is
+        # important for backward compatibility so miners dont lose incentive
+        try:
+            if (
+                synapse.stream_id in sent_preds
+                and synapse.dendrite.hotkey in sent_preds[synapse.stream_id]
+            ):
+                stream_preds = sent_preds[synapse.stream_id][synapse.dendrite.hotkey]
+                synapse.predictions = bt.tensor(np.array(stream_preds))
+            else:
+                bt.logging.warning(
+                    f"suspecting validator has not updated to V5, sending back non-hash based preds"
+                )
+                stream_preds = miner_preds[synapse.stream_id]
+                synapse.predictions = bt.tensor(np.array(stream_preds))
+            bt.logging.debug(f"sending lf [{stream_preds}]")
+            bt.logging.debug(f"sending lf with length [{len(stream_preds)}]")
+            return synapse
+        except Exception as e:
+            bt.logging.error(f"error returning synapse to vali: {e}")
+
+    # def lb_blacklist_fn(synapse: template.protocol.LiveBackward) -> tuple[bool, str]:
     #     # standardizing not accepting lb for now. Miner can override if they'd like.
     #     return False, synapse.dendrite.hotkey
     #
@@ -337,15 +398,17 @@ def main( config ):
     #     synapse.received = True
     #     return synapse
 
-    # Step 5: Build and link miner functions to the axon.
+    # Build and link miner functions to the axon.
     # The axon handles request processing, allowing validators to send this process requests.
     bt.logging.info(f"setting port [{config.axon.port}]")
     bt.logging.info(f"setting external port [{config.axon.external_port}]")
-    axon = bt.axon( wallet = wallet, port=config.axon.port, external_port=config.axon.external_port)
+    axon = bt.axon(
+        wallet=wallet, port=config.axon.port, external_port=config.axon.external_port
+    )
     bt.logging.info(f"Axon {axon}")
 
     # Attach determiners which functions are called when servicing a request.
-    bt.logging.info(f"Attaching forward function to axon.")
+    bt.logging.info(f"Attaching live forward functions to axon.")
     # axon.attach(
     #     forward_fn = training_f,
     #     blacklist_fn = tf_blacklist_fn,
@@ -357,12 +420,18 @@ def main( config ):
     #     priority_fn = tb_priority_fn,
     # )
     axon.attach(
-        forward_fn = live_f,
-        blacklist_fn = lf_blacklist_fn,
-        priority_fn = lf_priority_fn,
+        forward_fn=live_f,
+        blacklist_fn=lf_blacklist_fn,
+        priority_fn=lf_priority_fn,
+    )
+    axon.attach(
+        forward_fn=live_hash_f,
+        blacklist_fn=lf_hash_blacklist_fn,
+        priority_fn=lf_hash_priority_fn,
     )
 
-    # will leave live backward running from valis in case as a miner you'd like to use. Left out for majority of miners.
+    # Will leave live backward running from valis in case as a miner you'd like to use.
+    # Left out for majority of miners.
 
     # axon.attach(
     #     forward_fn = live_b,
@@ -372,31 +441,43 @@ def main( config ):
 
     # Serve passes the axon information to the network + netuid we are hosting on.
     # This will auto-update if the axon port of external ip have changed.
-    bt.logging.info(f"Serving attached axons on network:"
-                    f" {config.subtensor.chain_endpoint} with netuid: {config.netuid}")
-    axon.serve(netuid = config.netuid, subtensor = subtensor )
+    bt.logging.info(
+        f"Serving attached axons on network:"
+        f" {config.subtensor.chain_endpoint} with netuid: {config.netuid}"
+    )
+    axon.serve(netuid=config.netuid, subtensor=subtensor)
 
     # Start  starts the miner's axon, making it active on the network.
     bt.logging.info(f"Starting axon server on port: {config.axon.port}")
     axon.start()
 
-    # Step 6: Keep the miner alive
+    # Keep the miner alive
     # This loop maintains the miner's operations until intentionally stopped.
     bt.logging.info(f"Starting main loop")
     step = 0
 
     if base_mining_model is None:
-        bt.logging.error(f"base model not chosen, please pass using --base_model input arg")
+        bt.logging.error(
+            f"base model not chosen, please pass using --base_model input arg"
+        )
         sys.exit(0)
 
-    stream_predictions = [StreamPrediction.init_stream_prediction(request_template)
-                          for request_template in RequestTemplates().templates]
+    feature_collector = FeatureCollector(
+        sources=model_feature_sources,
+        feature_ids=model_feature_ids,
+        cache_results=True,
+        timeout=_FEATURE_COLLECTOR_TIMEOUT_MS,
+    )
 
-    data_generator_handler = DataGeneratorHandler()
+    stream_predictions = [
+        StreamPrediction.init_stream_prediction(request_template)
+        for request_template in RequestTemplates().templates
+    ]
 
-    run_update_predictions = threading.Thread(target=update_predictions, args=(stream_predictions,
-                                                                        data_generator_handler,
-                                                                        model_chosen))
+    run_update_predictions = threading.Thread(
+        target=update_predictions,
+        args=(stream_predictions, feature_collector, model_chosen),
+    )
     run_update_predictions.start()
 
     while True:
@@ -404,14 +485,16 @@ def main( config ):
             # Below: Periodically update our knowledge of the network graph.
             if step % 5 == 0:
                 metagraph = subtensor.metagraph(config.netuid)
-                log =  (f'Step:{step} | '\
-                        f'Block:{metagraph.block.item()} | '\
-                        f'Stake:{metagraph.S[my_subnet_uid]} | '\
-                        f'Rank:{metagraph.R[my_subnet_uid]} | '\
-                        f'Trust:{metagraph.T[my_subnet_uid]} | '\
-                        f'Consensus:{metagraph.C[my_subnet_uid] } | '\
-                        f'Incentive:{metagraph.I[my_subnet_uid]} | '\
-                        f'Emission:{metagraph.E[my_subnet_uid]}')
+                log = (
+                    f"Step:{step} | "
+                    f"Block:{metagraph.block.item()} | "
+                    f"Stake:{metagraph.S[my_subnet_uid]} | "
+                    f"Rank:{metagraph.R[my_subnet_uid]} | "
+                    f"Trust:{metagraph.T[my_subnet_uid]} | "
+                    f"Consensus:{metagraph.C[my_subnet_uid] } | "
+                    f"Incentive:{metagraph.I[my_subnet_uid]} | "
+                    f"Emission:{metagraph.E[my_subnet_uid]}"
+                )
                 bt.logging.info(log)
             step += 1
             time.sleep(1)
@@ -419,10 +502,11 @@ def main( config ):
         # If someone intentionally stops the miner, it'll safely terminate operations.
         except KeyboardInterrupt:
             axon.stop()
-            bt.logging.success('Miner killed by keyboard interrupt.')
+            bt.logging.success("Miner killed by keyboard interrupt.")
             break
+
         # In case of unforeseen errors, the miner will log the error and continue operations.
-        except Exception as e:
+        except Exception as main_loop_exception:  # noqa
             bt.logging.error(traceback.format_exc())
             continue
 
@@ -431,4 +515,4 @@ def main( config ):
 
 # This is the main function, which runs the miner.
 if __name__ == "__main__":
-    main( get_config() )
+    main(get_config())
