+++ vali_objects/scoring/scoring.py
@@ -5,16 +5,18 @@ import math
 from typing import List, Tuple, Dict
 
 import numpy as np
+from numpy import ndarray
 
 from vali_config import ValiConfig
-from vali_objects.exceptions.incorrect_prediction_size_error import IncorrectPredictionSizeError
+from vali_objects.exceptions.incorrect_prediction_size_error import (
+    IncorrectPredictionSizeError,
+)
 from vali_objects.exceptions.min_responses_exception import MinResponsesException
 from vali_objects.utils.vali_bkp_utils import ValiBkpUtils
 from vali_objects.utils.vali_utils import ValiUtils
 
 
 class Scoring:
-
     @staticmethod
     def calculate_weighted_rmse(predictions: np, actual: np) -> float:
         predictions = np.array(predictions)
@@ -32,21 +34,25 @@ class Scoring:
     def calculate_directional_accuracy(predictions: np, actual: np) -> float:
         pred_len = len(predictions)
 
-        pred_dir = np.sign([predictions[i] - predictions[i - 1] for i in range(1, pred_len)])
+        pred_dir = np.sign(
+            [predictions[i] - predictions[i - 1] for i in range(1, pred_len)]
+        )
         actual_dir = np.sign([actual[i] - actual[i - 1] for i in range(1, pred_len)])
 
         correct_directions = 0
-        for i in range(0, pred_len-1):
+        for i in range(0, pred_len - 1):
             correct_directions += actual_dir[i] == pred_dir[i]
 
-        return correct_directions / (pred_len-1)
+        return correct_directions / (pred_len - 1)
 
     @staticmethod
     def score_response(predictions: np, actual: np) -> float:
         if len(predictions) != len(actual) or len(predictions) == 0 or len(actual) < 2:
-            raise IncorrectPredictionSizeError(f"the number of predictions or the number of responses "
-                                               f"needed are incorrect: preds: '{len(predictions)}',"
-                                               f" results: '{len(actual)}'")
+            raise IncorrectPredictionSizeError(
+                f"the number of predictions or the number of responses "
+                f"needed are incorrect: preds: '{len(predictions)}',"
+                f" results: '{len(actual)}'"
+            )
 
         rmse = Scoring.calculate_weighted_rmse(predictions, actual)
 
@@ -71,10 +77,16 @@ class Scoring:
         min_score = min(score for _, score in scores)
         max_score = max(score for _, score in scores)
 
-        normalized_scores = [(name, (score - min_score) / (max_score - min_score)) for name, score in scores]
+        normalized_scores = [
+            (name, (score - min_score) / (max_score - min_score))
+            for name, score in scores
+        ]
         total_normalized_score = sum(score for _, score in normalized_scores)
 
-        normalized_scores = [(name, round(score / total_normalized_score, 4)) for name, score in normalized_scores]
+        normalized_scores = [
+            (name, round(score / total_normalized_score, 4))
+            for name, score in normalized_scores
+        ]
 
         return normalized_scores
 
@@ -86,7 +98,10 @@ class Scoring:
         min_score = min(score_values)
         max_score = max(score_values)
 
-        return {miner_uid:  1 - ((score - min_score) / (max_score - min_score)) for miner_uid, score in scores.items()}
+        return {
+            miner_uid: 1 - ((score - min_score) / (max_score - min_score))
+            for miner_uid, score in scores.items()
+        }
 
     @staticmethod
     def history_of_values() -> None | Dict[str, float]:
@@ -102,21 +117,26 @@ class Scoring:
         return 1
 
     @staticmethod
-    def get_geometric_mean_of_percentile(ds: List[List[float]]):
+    def get_geometric_mean_of_percentile(ds: ndarray):
         min_max_ranges_percentiled = ValiConfig.MIN_MAX_RANGES_PERCENTILED
         std_dev_ranges_percentiled = ValiConfig.STD_DEV_RANGES_PERCENTILED
 
-        results_min_max = max(ds[1]) / min(ds[1])
-        results_std_dev = np.std(ds[1])
+        results_min_max = max(ds) / min(ds)
+        results_std_dev = np.std(ds)
 
-        min_max_percentile = Scoring.get_percentile(results_min_max, min_max_ranges_percentiled)
-        std_dev_percentile = Scoring.get_percentile(results_std_dev, std_dev_ranges_percentiled)
+        min_max_percentile = Scoring.get_percentile(
+            results_min_max, min_max_ranges_percentiled
+        )
+        std_dev_percentile = Scoring.get_percentile(
+            results_std_dev, std_dev_ranges_percentiled
+        )
 
         return math.sqrt(min_max_percentile * std_dev_percentile)
 
     @staticmethod
-    def update_weights_using_historical_distributions(scores: List[Tuple[str, float]], ds: List[List[float]]):
-
+    def update_weights_using_historical_distributions(
+        scores: List[Tuple[str, float]], ds: ndarray
+    ):
         vweights = ValiUtils.get_vali_weights_json()
         geometric_mean_of_percentile = Scoring.get_geometric_mean_of_percentile(ds)
 
@@ -129,18 +149,22 @@ class Scoring:
 
         for key, value in vweights.items():
             if key not in score_miner_uids:
-                vweights[key] = Scoring.basic_ema((vweights[key] +
-                                                        (0 * geometric_mean_of_percentile))
-                                                       / (1 + geometric_mean_of_percentile), vweights[key])
+                vweights[key] = Scoring.basic_ema(
+                    (vweights[key] + (0 * geometric_mean_of_percentile))
+                    / (1 + geometric_mean_of_percentile),
+                    vweights[key],
+                )
 
         for score in scores:
             if score[0] in vweights:
                 previous_ema = vweights[score[0]]
             else:
                 previous_ema = vweight_avg
-            vweights[score[0]] = Scoring.basic_ema((previous_ema +
-                                                    (score[1] * geometric_mean_of_percentile))
-                                                   / (1 + geometric_mean_of_percentile), previous_ema)
+            vweights[score[0]] = Scoring.basic_ema(
+                (previous_ema + (score[1] * geometric_mean_of_percentile))
+                / (1 + geometric_mean_of_percentile),
+                previous_ema,
+            )
 
         ValiUtils.set_vali_weights_bkp(vweights)
         return vweights, geometric_mean_of_percentile
